Bootstrap params modified, utility func moved to back tools
